requirements.txt
numpy==2.1.2
torch==2.9.0
torchvision==0.24.0
matplotlib==3.10.0
streamlit==1.39.0
pillow==10.4.0